Revert "Illustrate moving the flags from the scope to the integration"
This reverts commit 430387fbbdc996328cce0f177fc537df88534ced.
